+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project (fork) adheres to [Semantic Versioning](http://semver.org/spec/
 since version 4.1.0.
 
 ## [Unreleased]
+### Added
+- support for customizing RakPeer UDP socket buffer sizes
 
 ## [4.2.0] - 2018-09-06
 ### Added